create menuhelper.rs for all the useful menu things
